Create Oct-2018-Goals.md
Example of Initial Goals Document. Copy the template, create a file for each month. Start editing this file at the beginning of each month cycle keeping the PR open to capture any back and forth dialogue between manager and employee. At the end of the month, close this PR out, and begin the next month PR.

At the beginning of the month, edit the 'Career Trajectory' section, as well as the 'This Month's Goals' section.
At the end of the month, use the 'Assessment of goal'. 100% success is not always the goal, we want to understand when we fell short, and how we can approach something differently next time. Maybe your manager needs to help you differently, or we want to adjust the goal to accomplish something more effective.

Each month the manager should do a lightweight review, and share that feedback early and often with employees. The idea is to work on issues early, before they become harder to solve. We want to be supportive and proactive to help the employee do their best! We also want to be transparent, so there is never any doubt or wasted mental energy worrying about "What does my manager think of me?".